Update - Prod use changes
Reduce the amount of config values
Remove logging:
  * app
  * environment
  * format
  * req_headers
  * server
  * state
These fields were either not needed or duplicated by filebeat logging
the hostname which contains the env & app in its name.

Added Logging:
  * fastly_duration
  * user_agent

Increase the value size to 500 chars as 50% of our valid requests were
being cut off.

Bump the version to 0.1.0 as it is now used in production on high
traffic services.