Reduce the number of strings allocated
* Replace interpolating two atoms converted to two strings and then
joined to create a third string. Instead convert the atoms to charlists
and convert the final joined charlist to a binary string

* Rework zero pad to recursively prepend zeros to an integer list until
the desired size is reached. Then convert the list to chars that are
joined with other data chars and then finally converted into a list.
This gets rid of 11 string joins.